back/requirements.txt
fastapi==0.109.0
psycopg2_binary==2.9.9
uvicorn==0.27.0
websockets
